Implement Quarters Feature
Add quarter fields to actions on db.
Add logic to track the current quarter and upload actions to db with current quarter.
Add logic and layout to match history to display actions with quarter information in actions record,
Add logic and layout to match history to display a score, goal, and behind breakdown by quarter and team.
Comment the new code.