Create 2 teams and a match with prefilled data for debugging
Added two JSON files containing prefilled data.
Added code to import JSON data and upload to firestore in ActivityMain.

NOTE: Both the JSONs and code for this was created by Copilot (M365 and in the IDE). I feel this is justified as none of it will be considered when marking and is merely for testing and debugging.