Completely reworked new match activity
Alot of code was redundant and therefore removed.
You can no longer add and create new players and teams on this screen, that can now only be done in team management.
The activity now allows you to select two already made teams from the db and displays the players on each team.